User reworked

+++ database/migrations/2026_05_02_175831_create_wallets_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->foreignId('user_id')->constrained()->cascadeOnDelete()->unique();
-            $table->decimal('balance')->default(0);
+            $table->bigInteger('balance')->default(0);
             $table->enum('currency', ['SYP', 'USD']);
 
             $table->timestamps();
